refactor: restructored template folders
