+++ packages/create-node-app-generator/README.md
@@ -1,1 +1,13 @@
 # create-node-app-generator
+
+## Install
+
+```bash
+$ npm i -g @reapit/create-node-app-generator
+```
+
+## Usage
+
+```bash
+$ 
+```
