+++ packages/create-node-app-generator/app/templates/express/.eslintrc.js
