+++ packages/create-node-app-generator/app/templates/express/_README.md
@@ -1,0 +1,1 @@
+# <%= name %>
