+++ packages/create-node-app-generator/app/templates/express/_package.json
