+++ packages/create-node-app-generator/app/templates/express/babel.config.js
