+++ packages/create-node-app-generator/app/templates/express/cdk/_cdk.ts
