+++ packages/create-node-app-generator/app/templates/express/jest.config.js
