+++ packages/create-node-app-generator/app/templates/express/src/aws.ts
