+++ packages/create-node-app-generator/app/templates/express/src/functions/hello.ts
