+++ packages/create-node-app-generator/app/templates/express/src/functions/index.ts
