+++ packages/create-node-app-generator/app/templates/express/tsconfig.json
