--- packages/create-node-app-generator/apps/express/README.md
@@ -1,1 +1,0 @@
-# express
